requirements.txt
numpy==2.3.1
pillow==11.3.0
tritonclient==2.59.0
requests==2.32.4
requests-toolbelt==1.0.0